Issue #5 Se crea clase presupuesto
